perf(renderer): isolate Recharts from activity sparklines
